requirements.txt
requests
phonenumbers
customtkinter
pymediainfo
datetime
pdf2docx
yt_dlp
keyboard
uuid
moviepy
discord
pydub
pyautogui
argparse
docx
tqdm
browser_cookie3
pyttsx3
asyncio
colorama
selenium
screeninfo
bcrypt
pathlib
pyaudio
bs4
httpx
importlib
requests
pyaes
aiohttp
cryptography
PyQt5
pyperclip
requests_toolbelt
phonenumbers
customtkinter
pymediainfo
datetime
pdf2docx
yt_dlp
keyboard
uuid
moviepy
discord
pydub
argparse
docx
tqdm
browser_cookie3
psutil
asyncio
colorama
selenium
bcrypt
pathlib
bs4
httpx
importlib
aiohttp
PyQt5
pyperclip
Logger
opencv-python
pycountry
pyperclip
pypiwin32
requests-toolbelt
wmi
pyaesm
browser-cookie3
Builder
nuitka
packaging
pillow
pycryptodome
pycryptodomex
pyinstaller
pycaw

